requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-multipart==0.0.12
jinja2==3.1.4
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pydantic-settings==2.6.1
email-validator==2.2.0
orjson==3.10.11
alembic==1.13.3

